Split out extension methods into classes
